--- core/deployment/src/main/java/io/quarkiverse/operatorsdk/deployment/helm/HelmGenerationEnabled.java
@@ -1,14 +1,0 @@
-package io.quarkiverse.operatorsdk.deployment.helm;
-
-import java.util.function.BooleanSupplier;
-
-import io.quarkiverse.operatorsdk.runtime.BuildTimeOperatorConfiguration;
-
-class HelmGenerationEnabled implements BooleanSupplier {
-    private BuildTimeOperatorConfiguration config;
-
-    @Override
-    public boolean getAsBoolean() {
-        return config.helm().enabled();
-    }
-}
