--- core/deployment/src/main/resources/helm/static/README.md
@@ -1,10 +1,0 @@
-## Configuration
-
-The following values are configurable:
-
-- `watchNamespaces` - namespaces to be watched, either:
-  - a list of comma-separated namespace names
-  - `JOSDK_ALL_NAMESPACES` to watch all namespaces
-  - `JOSDK_WATCH_CURRENT` to watch only the namespace in which the operator is deployed
-- `version` - the current version of the application.
-
