--- core/deployment/src/main/resources/helm/static/generic-crd-cluster-role-binding.yaml
@@ -1,16 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: {{ .Chart.Name }}-crd-validating-role-binding
-  labels:
-    app.kubernetes.io/name: {{ .Chart.Name }}
-    app.kubernetes.io/version: {{ .Chart.AppVersion }}
-    app.kubernetes.io/managed-by: quarkus
-roleRef:
-  kind: ClusterRole
-  apiGroup: rbac.authorization.k8s.io
-  name: {{ .Chart.Name }}-crd-validating-cluster-role
-subjects:
-  - kind: ServiceAccount
-    name: {{ .Chart.Name }}
-    namespace: {{ .Release.Namespace }}
