--- core/deployment/src/main/resources/helm/static/values.schema.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft-07/schema#",
-  "title": "Values",
-  "type": "object",
-  "properties": {
-    "watchNamespaces": {
-      "type": "string"
-    }
-  }
-}
