+++ helm/deployment/src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,0 +1,1 @@
+io.quarkiverse.operatorsdk.deployment.helm.HelmOverridesConfigSource
