+++ helm/deployment/src/main/resources/helm/additional-crd-role-binding-template.yaml
@@ -1,0 +1,56 @@
+{{ if eq $.Values.watchNamespaces "JOSDK_WATCH_CURRENT" }}
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  name: {role-binding-name}
+  namespace: {{ $.Release.Namespace }}
+  labels:
+    app.kubernetes.io/name: {{ .Chart.Name }}
+    app.kubernetes.io/version: {{ .Chart.AppVersion }}
+    app.kubernetes.io/managed-by: quarkus
+roleRef:
+  kind: {role-ref-kind}
+  apiGroup: {role-ref-api-group}
+  name: {role-ref-name}
+subjects:
+  - kind: ServiceAccount
+    name: {{ $.Chart.Name }}
+{{ else if eq $.Values.watchNamespaces "JOSDK_ALL_NAMESPACES" }}
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: {role-binding-name}
+  labels:
+    app.kubernetes.io/name: {{ .Chart.Name }}
+    app.kubernetes.io/version: {{ .Chart.AppVersion }}
+    app.kubernetes.io/managed-by: quarkus
+roleRef:
+  kind: {role-ref-kind}
+  apiGroup: {role-ref-api-group}
+  name: {role-ref-name}
+subjects:
+  - kind: ServiceAccount
+    name: {{ $.Chart.Name }}
+    namespace: {{ $.Release.Namespace }}
+{{ else }}
+{{ range $anamespace := ( split "," $.Values.watchNamespaces ) }}
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  name: {role-binding-name}
+  namespace: {{ $anamespace }}
+  labels:
+    app.kubernetes.io/name: {{ .Chart.Name }}
+    app.kubernetes.io/version: {{ .Chart.AppVersion }}
+    app.kubernetes.io/managed-by: quarkus
+roleRef:
+  kind: {role-ref-kind}
+  apiGroup: {role-ref-api-group}
+  name: {role-ref-name}
+subjects:
+  - kind: ServiceAccount
+    name: {{ $.Chart.Name }}
+    namespace: {{ $.Release.Namespace }}
+---
+{{- end }}
+{{- end }}
