+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/InjectedDependency.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.operatorsdk.test.sources;
+
+import jakarta.inject.Singleton;
+
+@Singleton
+public class InjectedDependency {
+}
