+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/SimpleSpec.java
@@ -1,0 +1,5 @@
+package io.quarkiverse.operatorsdk.test.sources;
+
+public class SimpleSpec {
+    public String value;
+}
