+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/SimpleSpecV2.java
@@ -1,0 +1,6 @@
+package io.quarkiverse.operatorsdk.test.sources;
+
+public class SimpleSpecV2 {
+    public String value;
+    public int number;
+}
