+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/SimpleStatus.java
@@ -1,0 +1,5 @@
+package io.quarkiverse.operatorsdk.test.sources;
+
+public class SimpleStatus {
+    public String value;
+}
