+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/WatchAllNamespacesReconciler.java
