+++ helm/deployment/src/test/java/io/quarkiverse/operatorsdk/test/sources/WatchAllReconciler.java
@@ -1,0 +1,20 @@
+package io.quarkiverse.operatorsdk.test.sources;
+
+import io.fabric8.kubernetes.api.model.ConfigMap;
+import io.javaoperatorsdk.operator.api.config.informer.Informer;
+import io.javaoperatorsdk.operator.api.reconciler.Constants;
+import io.javaoperatorsdk.operator.api.reconciler.Context;
+import io.javaoperatorsdk.operator.api.reconciler.ControllerConfiguration;
+import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
+import io.javaoperatorsdk.operator.api.reconciler.UpdateControl;
+
+@ControllerConfiguration(name = WatchAllReconciler.NAME, informer = @Informer(namespaces = Constants.WATCH_ALL_NAMESPACES))
+public class WatchAllReconciler implements Reconciler<ConfigMap> {
+
+    public static final String NAME = "watchall";
+
+    @Override
+    public UpdateControl<ConfigMap> reconcile(ConfigMap configMap, Context<ConfigMap> context) throws Exception {
+        return UpdateControl.noUpdate();
+    }
+}
