+++ helm/deployment/src/test/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.http.test-port=0
+quarkus.operator-sdk.generate-with-watched-namespaces=operator-level-buildtime-ns
+quarkus.operator-sdk.controllers.simple.generate-with-watched-namespaces=simple-ns
