+++ helm/runtime/pom.xml
@@ -3,15 +3,15 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>io.quarkiverse.operatorsdk</groupId>
-    <artifactId>quarkus-operator-sdk-core-parent</artifactId>
+    <artifactId>quarkus-operator-sdk-helm-parent</artifactId>
     <version>7.2.1-SNAPSHOT</version>
   </parent>
-  <artifactId>quarkus-operator-sdk</artifactId>
-  <name>Quarkus - Operator SDK - Core - Runtime</name>
+  <artifactId>quarkus-operator-sdk-helm</artifactId>
+  <name>Quarkus - Operator SDK - Helm - Runtime</name>
   <dependencies>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-common</artifactId>
+      <artifactId>quarkus-operator-sdk</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
