+++ pom.xml
@@ -14,6 +14,7 @@
   <properties>
     <quarkus.version>3.25.0</quarkus.version>
     <java-operator-sdk.version>5.1.2</java-operator-sdk.version>
+    <quarkiverse.helm.version>1.2.7</quarkiverse.helm.version>
   </properties>
   <scm>
     <connection>:git:git@github.com:quarkiverse/quarkus-operator-sdk.git</connection>
@@ -65,6 +66,7 @@
         <module>common-deployment</module>
         <module>cli</module>
         <module>core</module>
+        <module>helm</module>
         <module>bundle-generator</module>
         <module>integration-tests</module>
       </modules>
